Refactor: Move config and run files to appropriate locations in infrastructure
