+++ src/infrastructure/api/run.py
@@ -1,7 +1,7 @@
 """アプリケーション起動スクリプト"""
 import uvicorn
 from src.infrastructure.api.main import app
-from config.api import API_HOST, API_PORT
+from src.infrastructure.api.config import API_HOST, API_PORT
 
 if __name__ == "__main__":
     uvicorn.run(
